PR feedback resolved
