+++ package.json
@@ -12,7 +12,6 @@
   "dependencies": {
     "lodash": "^4.17.21",
     "mapbox-gl": "^3.10.0",
-    "path": "^0.12.7",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "vite-tsconfig-paths": "^5.1.4"
