+++ src/App.tsx
@@ -1,9 +1,7 @@
-import { useState, useMemo, useCallback, useEffect } from "react";
-import { debounce } from "lodash";
+import { useState, useMemo } from "react";
 
 import Filters from "@components/Header/Filters";
-import MapView from "@components/ViewTabs/MapView";
-import ListView from "@components/ViewTabs/ListView";
+import ProjectsView from "@src/components/ProjectsView";
 import { IProject } from "@interfaces";
 import { ViewType } from "@constants";
 
@@ -16,19 +14,6 @@ function App() {
   const [searchQuery, setSearchQuery] = useState<string>("");
   const [selectedProject, setSelectedProject] = useState<IProject | null>(null);
   const [toggleView, setToggleView] = useState<string>(ViewType.MAP);
-
-  const [debouncedSearchQuery, setDebouncedSearchQuery] = useState(searchQuery);
-
-  // Debounce the search query update
-  const debouncedSetSearchQuery = useCallback(
-    debounce((query) => setDebouncedSearchQuery(query), 300),
-    []
-  );
-
-  // Update the debounced query when searchQuery changes
-  useEffect(() => {
-    debouncedSetSearchQuery(searchQuery);
-  }, [searchQuery, debouncedSetSearchQuery]);
 
   const filteredPins = useMemo(() => {
     return mapPins.filter((pin) => {
@@ -44,7 +29,7 @@ function App() {
 
       return matchCategory && matchLocation && matchSearch;
     }) as IProject[];
-  }, [selectedCategory, selectedLocation, debouncedSearchQuery]);
+  }, [selectedCategory, selectedLocation, searchQuery]);
 
   return (
     <div className="app-container">
@@ -55,7 +40,6 @@ function App() {
             setSelectedCategory={setSelectedCategory}
             selectedLocation={selectedLocation}
             setSelectedLocation={setSelectedLocation}
-            searchQuery={searchQuery}
             setSearchQuery={setSearchQuery}
             toggleView={toggleView}
             setToggleView={setToggleView}
@@ -68,18 +52,12 @@ function App() {
             justifyContent: "center",
           }}>
           <div style={{ width: "1200px" }}>
-            {toggleView === ViewType.MAP ? (
-              <MapView
-                pins={filteredPins}
-                selectedProject={selectedProject}
-                setSelectedProject={setSelectedProject}
-              />
-            ) : (
-              <ListView
-                pins={filteredPins}
-                setSelectedProject={setSelectedProject}
-              />
-            )}
+            <ProjectsView
+              pins={filteredPins}
+              selectedProject={selectedProject}
+              setSelectedProject={setSelectedProject}
+              isListView={toggleView === ViewType.LIST}
+            />
           </div>
         </div>
       </div>
