+++ src/components/ProjectsView/ProjectsList.tsx
@@ -5,22 +5,26 @@ import ProjectCard from "./ProjectCard";
 interface IProjectsList {
   pins: IProject[],
   setSelectedProject: (project: IProject) => void,
+  isListView: boolean
 };
 
 const ProjectsList: React.FC<IProjectsList> = ({
   pins,
   setSelectedProject,
+  isListView
 }) => {
   return (
-    <div className="locations-list">
+    <div className={!isListView ? "locations-list" : ""}>
       {pins.length ? (
-        pins.map((project) => (
-          <ProjectCard
-            key={project.id}
-            project={project}
-            setSelectedProject={setSelectedProject}
-          />
-        ))
+        <div className={isListView ? "grid-container" : ""}>
+          {pins.map((project) => (
+            <ProjectCard
+              key={project.id}
+              project={project}
+              setSelectedProject={setSelectedProject}
+            />
+          ))}
+        </div>
       ) : (
         <div className="no-project-text">No Project Found</div>
       )}
