+++ src/components/ProjectsView/index.tsx
@@ -1,0 +1,27 @@
+import { IProject } from "@interfaces";
+import MapComponent from "@components/Map";
+import ProjectsList from "./ProjectsList";
+
+interface IProjectsViewProps {
+  pins: IProject[];
+  selectedProject: IProject | null;
+  setSelectedProject: (project: IProject) => void;
+  isListView: boolean
+}
+
+const ProjectsView: React.FC<IProjectsViewProps> = ({ pins, selectedProject, setSelectedProject, isListView }) => {
+  return (
+    <div className={isListView ? "list-view-container" : "projects-container"}>
+      <ProjectsList pins={pins} setSelectedProject={setSelectedProject} isListView={isListView} />
+      <div className={`map-container ${isListView ? "hidden" : ""}`}>
+        <MapComponent
+          pins={pins}
+          selectedProject={selectedProject}
+          setSelectedProject={setSelectedProject}
+        />
+      </div>
+    </div>
+  );
+};
+
+export default ProjectsView;
