+++ vite.config.ts
@@ -1,6 +1,5 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
-import path from "path";
 import tsconfigPaths from "vite-tsconfig-paths";
 
 // https://vite.dev/config/
@@ -9,8 +8,5 @@ export default defineConfig({
   assetsInclude: ["**/*.jpg", "**/*.png", "**/*.jpeg", "**/*.svg"],
   server: {
     allowedHosts: ["1005-2a09-bac5-5041-16a0-00-241-6d.ngrok-free.app"],
-  },
-  resolve: {
-    alias: [{ find: "@", replacement: path.resolve(__dirname, "src") }],
   },
 });
